requirements.txt
aiogram==3.*
python-dotenv
fatsecret
rauth